Pair Project on Four Operations/requirements.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
coverage==7.6.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.39.3
idna==3.4
ipython==8.12.3
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
line_profiler==4.1.3
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.28.2
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
